+++ AlgoriaCore.WebUI/ClientApp/src/app/catalogscustomimpl/catalogscustomimpl.component.ts
@@ -251,17 +251,19 @@ export class CatalogsCustomImplComponent extends AppComponentBase implements OnI
 
         self.alertService.confirm(
             self.l('ConfirmationSoftDeleteRecordMessage'), self.l('Confirmation'),
-            function () {
-                self.app.blocked = true;
+            function (res) {
+                if (res === true) {
+                    self.app.blocked = true;
 
-                self.service.deleteCatalogCustomImpl(new CatalogCustomImplDeleteCommand({ id: id, catalog: self.catalogId }))
-                    .pipe(finalize(() => {
-                        self.app.blocked = false;
-                    }))
-                    .subscribe(data => {
-                        self.notify.success(self.l('RecordDeletedSuccessful'), self.l('Success'));
-                        self.getList();
-                    });
+                    self.service.deleteCatalogCustomImpl(new CatalogCustomImplDeleteCommand({ id: id, catalog: self.catalogId }))
+                        .pipe(finalize(() => {
+                            self.app.blocked = false;
+                        }))
+                        .subscribe(data => {
+                            self.notify.success(self.l('RecordDeletedSuccessful'), self.l('Success'));
+                            self.getList();
+                        });
+                }
             }
         );
     }
